requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
